Interfaces de Data para ID
