+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/external/TheMovieDataBase.kt
@@ -1,0 +1,56 @@
+package edu.dyds.movies.data.external
+
+import edu.dyds.movies.domain.entity.Movie
+import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.request.get
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+class TheMovieDataBase(
+    private val tmdbHttpClient: HttpClient,
+) : MoviesExternalSource {
+    override suspend fun getPopularMovies(): RemoteResult =
+        tmdbHttpClient.get("/3/discover/movie?sort_by=popularity.desc").body()
+
+    override suspend fun getMovieDetailsDB(id: Int): RemoteMovie =
+        tmdbHttpClient.get("/3/movie/$id").body()
+}
+
+@Serializable
+data class RemoteResult(
+    val page: Int,
+    val results: List<RemoteMovie>,
+    @SerialName("total_pages") val totalPages: Int,
+    @SerialName("total_results") val totalResults: Int
+)
+
+@Serializable
+data class RemoteMovie(
+    val id: Int,
+    val title: String,
+    val overview: String,
+    @SerialName("release_date") val releaseDate: String,
+    @SerialName("poster_path") val posterPath: String,
+    @SerialName("backdrop_path") val backdropPath: String?,
+    @SerialName("original_title") val originalTitle: String,
+    @SerialName("original_language") val originalLanguage: String,
+    val popularity: Double,
+    @SerialName("vote_average") val voteAverage: Double,
+
+    ) {
+    fun toDomainMovie(): Movie {
+        return Movie(
+            id = id,
+            title = title,
+            overview = overview,
+            releaseDate = releaseDate,
+            poster = "https://image.tmdb.org/t/p/w185$posterPath",
+            backdrop = backdropPath.let { "https://image.tmdb.org/t/p/w780$it" },
+            originalTitle = originalTitle,
+            originalLanguage = originalLanguage,
+            popularity = popularity,
+            voteAverage = voteAverage
+        )
+    }
+}
